resolve review comments (function relocation)
