+++ dds/DCPS/RTPS/BaseMessageUtils.h
@@ -227,6 +227,9 @@ private:
   size_t smContentStart_;
 };
 
+OpenDDS_Rtps_Export
+bool bitmapNonEmpty(const SequenceNumberSet& snSet);
+
 }
 }
