--- dds/DCPS/RTPS/Util.h
@@ -1,30 +1,0 @@
-/*
- *
- *
- * Distributed under the OpenDDS License.
- * See: http://www.opendds.org/license.html
- */
-
-#ifndef OPENDDS_RTPS_UTIL_H
-#define OPENDDS_RTPS_UTIL_H
-
-#if !defined (ACE_LACKS_PRAGMA_ONCE)
-#pragma once
-#endif /* ACE_LACKS_PRAGMA_ONCE */
-
-#include "rtps_export.h"
-#include "RtpsCoreC.h"
-
-OPENDDS_BEGIN_VERSIONED_NAMESPACE_DECL
-
-namespace OpenDDS {
-namespace RTPS {
-
-OpenDDS_Rtps_Export bool bitmapNonEmpty(const SequenceNumberSet& snSet);
-
-} // namespace RTPS
-} // namespace OpenDDS
-
-OPENDDS_END_VERSIONED_NAMESPACE_DECL
-
-#endif
